refactor(linter): add output formatter
