+++ crates/oxc_linter/src/output_formatter/json.rs
